requirements.txt
python-dotenv~=1.0.1
langchain-openai
langchain-text-splitters
lxml
bs4
langchain_community
unstructured[html]
psutil
chromadb